requirements.txt
django==3.2.18
djangorestframework==3.14.0
djoser==2.1.0
gunicorn==20.1.0
whitenoise==6.3.0
django-cors-headers==3.14.0
django-currentuser==0.5.3